+++ src-tauri/src/remote_service/transport.rs
@@ -207,6 +207,10 @@ pub async fn connected(
         "x-nexus-application-keyboard-version",
         "1".parse().map_err(|_| "invalidResponse")?,
     );
+    request.headers_mut().insert(
+        "x-nexus-application-insights-version",
+        "1".parse().map_err(|_| "invalidResponse")?,
+    );
     let config = WebSocketConfig::default()
         .max_message_size(Some(512))
         .max_frame_size(Some(512))